Finished ringbuffered upload implementation
XL3 data is now buffered, and threading works correctly, too.
Instead of taking the time to do a full copy to allow the threading
to work without mutexes, I wrote a ringbuf_copy function which
just creates a new Ringbuf which is given all of the pointers that
the original had. All that is copied are void pointers, so it's
pretty fast. Upload is still fast (to a local machine, anyway).

I also fixed a bug in ringbuf_push - only the first size(input_ptr)
bytes were being copied. Now, the function accepts a size argument,
which allows everything to be copied correctly.