First version of snotstream with pouch and libcurl
I modified the pouchlib included in this project
to only have a multi interface, and no document
or database wrapper functions. So it's back to low
level. But now, snotstream takes a url in from a
controller terminal and creates a pouch request,
which goes through the multi interface of libcurl
and performs a GET to whatever the url is. It's
simple right now, but it proves that the new
pouch interface works, so I can start working on
actually doing DB requests next.